Minor fix (#20)
* Make customQueriesPath config an optional config

* Issue: User has to cd or ls to customQueries directory before the user can access the file.
Cause: This is because the files will not be populate until the directory is opened (using cd or ls)
Fix: Need to populate all files in customqueries dir when DBFS start up. This is so that the user can access the file directly with full path name.

Also refactor files creation code.